requirements.txt
# ===================================================================
# ALPR Thesis Project - Python Dependencies
# ===================================================================
# Compatible with: Python 3.8+, Linux/macOS/Windows
# 
# Linux Prerequisites (Ubuntu/Debian):
#   sudo apt-get install -y python3-pip python3-venv
#   sudo apt-get install -y libgl1-mesa-glx libglib2.0-0 libsm6 libxext6
#   sudo apt-get install -y build-essential python3-dev  # For pycocotools
#
# For GPU support (NVIDIA):
#   Ensure CUDA 11.8+ and compatible drivers are installed
#   Visit: https://pytorch.org/get-started/locally/
# ===================================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python              # Requires: libgl1-mesa-glx, libglib2.0-0 (Linux)
pillow                     # Image processing

# Data Processing
pyyaml                     # Configuration files
tqdm                       # Progress bars
scipy                      # Scientific computing
pandas                     # Data manipulation
numpy                      # Array operations

# Object Detection Models
ultralytics                # YOLOv11, YOLOv26 (unified framework)
rfdetr                     # RF-DETR-S (Roboflow detector)

# Additional Tools
onnx                       # Model export format
onnxruntime                # ONNX inference engine
tensorboard                # Training visualization
# pycocotools              # COCO tools (not required; project trains on YOLO format)

# Optional: Uncomment for additional features
# matplotlib                # Plotting and visualization
# seaborn                   # Statistical visualization
# jupyter                   # Interactive notebooks
# ipython                   # Enhanced Python shell